refactor: change file structure, delete old files
