+++ src/lib/components/Charts/ComboChart.svelte
@@ -1,212 +1,150 @@
-<script>
+<script lang="ts">
 	import PriceLegendIcon from '$lib/icons/PriceLegendIcon.svelte';
-	import { onMount } from 'svelte';
+	import ChartToolbar from './ChartToolbar.svelte';
 
-	let chartElement;
-	let innerWidth;
+	let ctx: HTMLCanvasElement;
 	let chart;
-	export let priceData = [];
-	export let volumeData = [];
+
+	export let priceData: [{ x?: Date; y?: number }];
+	export let volumeData: [{ x?: Date; y?: number }];
 	export let isLoading = true;
-	onMount(() => {
-		let options = {
-			chart: {
-				type: 'line',
-				toolbar: {
-					show: true,
-					offsetY: -12
-				},
-				zoom: {
-					enabled: true,
-					zoomedArea: {
-						fill: {
-							color: '#099B91',
-							opacity: 0.4
-						},
-						stroke: {
-							color: '#099B91',
-							opacity: 0.4,
-							width: 1
-						}
-					}
-				},
-				height: '100%',
-				width: `${innerWidth * 0.75}px`
-			},
-			stroke: {
-				curve: 'stepline',
-				width: [2, 0]
-			},
-			series: [
-				{
-					name: 'Price',
-					data: priceData,
-					type: 'line'
-				},
-				{
-					name: 'Volume',
-					data: volumeData,
-					type: 'column'
-				}
-			],
-			xaxis: {
-				type: 'datetime',
-				axisBorder: {
-					show: true
-				},
-				axisTicks: {
-					show: true
-				},
-				labels: {
-					datetimeFormatter: {
-						year: 'yyyy',
-						month: 'MMM yy',
-						day: 'dd MMM',
-						hour: 'HH:mm'
-					},
-					style: {
-						fontSize: '0.83vw',
-						colors: '#8F9398'
-					}
-				},
-				tickAmount: 3
-			},
-			yaxis: [
-				{
-					labels: {
-						formatter: (value) => {
-							if (value > 1000) {
-								return `$${Math.round(value / 1000).toFixed(3)}k`;
-							}
-							return `$${value.toFixed(3)}`;
-						},
-						style: {
-							fontSize: '0.83vw',
-							colors: ['#8F9398']
-						}
-					},
-					tickAmount: 5
-				},
-				{
-					labels: {
-						formatter: (value) => {
-							if (value > 1000) {
-								return `${Math.round(value / 1000)}k`;
-							}
-						},
-						style: {
-							fontSize: '0.83vw',
-							colors: ['#8F9398']
-						}
-					},
-					tickAmount: 5,
-					opposite: true
-				}
-			],
-			legend: {
-				show: false
-			},
-			colors: ['#099B91', '#0021A5'],
-			tooltip: {
-				enabled: true,
-				style: {
-					fontSize: '0.83vw'
-				},
-				x: {
-					show: false,
-					format: 'dddd, d MMM, HH:mm'
-				},
-				y: {
-					formatter: (value) => {
-						return value.toLocaleString();
-					}
-				},
-				custom: function ({ series, seriesIndex, dataPointIndex, w }) {
-					const monthNames = [
-						'January',
-						'February',
-						'March',
-						'April',
-						'May',
-						'June',
-						'July',
-						'August',
-						'September',
-						'October',
-						'November',
-						'December'
-					];
-					const date = new Date(w.globals.seriesX[seriesIndex][dataPointIndex]);
 
-					return (
-						'<div style="padding: clamp(4px, 0.83vw, 0.83vw); font-size: clamp(10px,0.83vw,0.83vw)">' +
-						'<div style="font-weight: bold;">' +
-						String(date.getDate()).padStart(2, '0') +
-						' ' +
-						monthNames[date.getMonth()] +
-						'</div>' +
-						'<div style="display: flex; gap: clamp(8px,0.83vw,0.83vw); justify-content: space-between">' +
-						'<span style="display: flex; align-items: center;">' +
-						'<div style="border-radius: 100%; height: clamp(8px,0.6vw,0.6vw); width: clamp(8px,0.6vw,0.6vw); margin-right: clamp(4px,0.24vw,0.24vw); background-color:' +
-						w.globals.colors[0] +
-						';"></div>' +
-						w.globals.initialSeries[0].name +
-						' : ' +
-						'</span>' +
-						'<span style="font-weight: bold;"> $' +
-						series[0][dataPointIndex].toLocaleString() +
-						'</span>' +
-						'</div>' +
-						'<div style="display: flex; gap: clamp(8px,0.83vw,0.83vw); justify-content: space-between">' +
-						'<span style="display: flex; align-items: center;">' +
-						'<div style="border-radius: 100%; height: clamp(8px,0.6vw,0.6vw); width: clamp(8px,0.6vw,0.6vw); margin-right: clamp(4px,0.24vw,0.24vw); background-color:' +
-						w.globals.colors[1] +
-						';"></div>' +
-						w.globals.initialSeries[1].name +
-						' : ' +
-						'</span>' +
-						'<span style="font-weight: bold;">' +
-						series[1][dataPointIndex].toLocaleString() +
-						'</span>' +
-						'</div>' +
-						'</div>'
-					);
-				}
-			},
-			grid: {
-				xaxis: {
-					lines: {
-						show: false
-					}
-				},
-				yaxis: {
-					lines: {
-						show: true
-					}
-				}
-			}
-		};
-		// @ts-ignore
-		chart = new ApexCharts(chartElement, options);
-		chart.render();
-	});
 	$: if (!isLoading) {
-		chart?.updateSeries([
-			{
-				name: 'Price',
-				data: priceData,
-				type: 'line'
-			},
-			{
-				name: 'Volume',
-				data: volumeData,
-				type: 'column'
-			}
-		]);
+		priceData?.length > 0 && volumeData?.length > 0 && renderChart(volumeData, priceData);
 	}
-</script>
 
-<svelte:window bind:innerWidth />
+	const renderChart = (
+		chartData1: [{ x?: Date; y?: number }],
+		chartData2: [{ x?: Date; y?: number }]
+	) => {
+		chart = new Chart(ctx, {
+			type: 'line',
+			data: {
+				datasets: [
+					{
+						label: 'Volume',
+						data: chartData1,
+						backgroundColor: '#0021A5',
+						borderColor: '#0021A5',
+						type: 'bar',
+						borderWidth: 2,
+						order: 1,
+						yAxisID: 'y'
+					},
+					{
+						label: 'Prices',
+						data: chartData2,
+						backgroundColor: '#099B91',
+						borderColor: '#099B91',
+						borderWidth: 2,
+						order: 0,
+						yAxisID: 'y1',
+						stepped: true,
+						pointStyle: 'circle',
+						pointRadius: 0
+					}
+				]
+			},
+			options: {
+				responsive: true,
+				interaction: {
+					mode: 'index',
+					intersect: false
+				},
+				scales: {
+					x: {
+						adapters: {
+							date: {}
+						},
+						type: 'time',
+						time: {
+							unit: 'day'
+						},
+						grid: {
+							display: false
+						},
+						ticks: {
+							autoSkip: true,
+							maxTicksLimit: 20,
+							maxRotation: 0,
+							minRotation: 0
+						}
+					},
+					y: {
+						type: 'linear',
+						display: true,
+						position: 'right',
+						beginAtZero: false,
+						grid: {
+							drawOnChartArea: true
+						},
+						ticks: {
+							callback: function (val, index) {
+								const lookup = [
+									{ value: 1, symbol: "" },
+									{ value: 1e3, symbol: "k" },
+									{ value: 1e6, symbol: "M" },
+									{ value: 1e9, symbol: "G" },
+									{ value: 1e12, symbol: "T" },
+									{ value: 1e15, symbol: "P" },
+									{ value: 1e18, symbol: "E" }
+								];
+								const rx = /\.0+$|(\.[0-9]*[1-9])0+$/;
+								var item = lookup.slice().reverse().find(function(item) {
+									return val >= item.value;
+								});
+								return item ? (val / item.value).toFixed(2).replace(rx, "$1") + item.symbol : "0";
+							}
+						}
+					},
+					y1: {
+						type: 'linear',
+						display: true,
+						position: 'left',
+						beginAtZero: false,
+						grid: {
+							drawOnChartArea: true
+						},
+						ticks: {
+							callback: function (val, index) {
+								return `$${val.toFixed(3)}`;
+							}
+						}
+					}
+				},
+				plugins: {
+					legend: {
+						display: false
+					},
+					tooltip: {
+						enabled: false,
+						position: 'nearest'
+						// external: externalTooltipHandler
+					},
+					zoom: {
+						pan: {
+							enabled: true,
+							mode: 'x',
+							threshold: 5
+						},
+						zoom: {
+							wheel: {
+								enabled: true
+							},
+							drag: {
+								enabled: false
+							},
+							pinch: {
+								enabled: true
+							},
+							mode: 'x'
+						}
+					}
+				}
+			}
+		});
+	};
+</script>
 
 <div class="container">
 	<div class="title">Market Price (CSPR/USD)</div>
@@ -222,7 +160,10 @@
 			<div class="text">Volume</div>
 		</div>
 	</div>
-	<div class="chart" bind:this={chartElement} />
+	<ChartToolbar {chart}/>
+	<div class="chart">
+		<canvas bind:this={ctx} />
+	</div>
 </div>
 
 <style lang="postcss">
@@ -232,9 +173,13 @@
 	}
 
 	.container {
-		@apply md:h-[32vw] min-w-max;
+		@apply min-w-max;
 		@apply flex flex-col items-center justify-center;
 		@apply my-[clamp(16px,0.95vw,0.95vw)];
+	}
+
+	.chart {
+		@apply w-full;
 	}
 
 	.legend {
