+++ src/lib/components/Charts/PoNegAreaChart.svelte
@@ -1,211 +1,115 @@
-<script>
-	import { onMount } from 'svelte';
-	let chartElement;
-	let innerWidth;
+<script lang="ts">
+	import ChartToolbar from "./ChartToolbar.svelte";
+
+	let ctx;
 	let chart;
-	export let delegatedData = [];
-	export let unbondedData = [];
+
+	export let delegatedData: [{ x?: Date; y?: number }];
+	export let unbondedData: [{ x?: Date; y?: number }];
 	export let isLoading = true;
-	onMount(() => {
-		// console.log(`${innerWidth * 0.75}px`)
-		let options = {
-			chart: {
-				type: 'area',
-				toolbar: {
-					show: true,
-					offsetY: -12
+
+	$: if (!isLoading) {
+		delegatedData?.length > 0 &&
+			unbondedData?.length > 0 &&
+			renderChart(delegatedData, unbondedData);
+	}
+
+	const renderChart = (
+		chartData1: [{ x?: Date; y?: number }],
+		chartData2: [{ x?: Date; y?: number }]
+	) => {
+		chart = new Chart(ctx, {
+			type: 'line',
+			data: {
+				datasets: [
+					{
+						label: 'Delegated',
+						data: chartData1,
+						backgroundColor: '#0021A5',
+						borderWidth: 0,
+						fill: 'origin',
+						pointStyle: 'circle',
+						pointRadius: 0
+					},
+					{
+						label: 'Unbonded',
+						data: chartData2,
+						backgroundColor: '#099B91',
+						borderWidth: 0,
+						fill: 'origin',
+						pointStyle: 'circle',
+						pointRadius: 0
+					}
+				]
+			},
+			options: {
+				responsive: true,
+				interaction: {
+					mode: 'index',
+					intersect: false
 				},
-				zoom: {
-					enabled: true,
-					zoomedArea: {
-						fill: {
-							color: '#099B91',
-							opacity: 0.4
+				scales: {
+					x: {
+						adapters: {
+							date: {}
 						},
-						stroke: {
-							color: '#099B91',
-							opacity: 0.4,
-							width: 1
+						type: 'time',
+						time: {
+							unit: 'day'
+						},
+						grid: {
+							display: false
+						},
+						ticks: {
+							autoSkip: true,
+							maxTicksLimit: 20,
+							maxRotation: 0,
+							minRotation: 0
+						}
+					},
+					y: {
+						type: 'linear',
+						display: true,
+						position: 'left',
+						beginAtZero: false,
+						grid: {
+							drawOnChartArea: true
 						}
 					}
 				},
-				height: '100%',
-				width: `${innerWidth * 0.75}px`
-			},
-			fill: {
-				type: 'solid'
-			},
-			dataLabels: {
-				enabled: false
-			},
-			stroke: {
-				curve: 'straight',
-				width: 0
-			},
-			legend: {
-				show: false
-			},
-			series: [
-				{
-					name: 'Delegated',
-					data: delegatedData,
-					color: '#0021A5'
-				},
-				{
-					name: 'Unbonded',
-					data: unbondedData,
-					color: '#099B91'
-				}
-			],
-			xaxis: {
-				type: 'datetime',
-				axisBorder: {
-					show: true
-				},
-				axisTicks: {
-					show: false
-				},
-				labels: {
-					formatter: (value) => {
-						const monthNames = [
-							'January',
-							'February',
-							'March',
-							'April',
-							'May',
-							'June',
-							'July',
-							'August',
-							'September',
-							'October',
-							'November',
-							'December'
-						];
-						let date = new Date(value);
-						return `${date.getDate()} ${monthNames[date.getMonth()]}`;
+				plugins: {
+					legend: {
+						display: false
 					},
-					style: {
-						fontSize: '0.83vw',
-						colors: '#CFCFCF'
+					tooltip: {
+						enabled: false,
+						position: 'nearest'
+						// external: externalTooltipHandler
 					},
-					offsetX: 8
-				}
-			},
-			yaxis: {
-				labels: {
-					formatter: (value) => {
-						return value.toLocaleString();
-					},
-					style: {
-						fontSize: '0.83vw',
-						colors: ['#8F9398']
-					},
-					offsetX: -16
-				},
-				tickAmount: 6
-			},
-			tooltip: {
-				enabled: true,
-				style: {
-					fontSize: '0.83vw'
-				},
-				x: {
-					show: false,
-					format: 'dddd, d MMM, HH:mm'
-				},
-				y: {
-					formatter: (value) => {
-						return value.toLocaleString();
-					}
-				},
-				custom: function ({ series, seriesIndex, dataPointIndex, w }) {
-					const monthNames = [
-						'January',
-						'February',
-						'March',
-						'April',
-						'May',
-						'June',
-						'July',
-						'August',
-						'September',
-						'October',
-						'November',
-						'December'
-					];
-					const date = new Date(w.globals.seriesX[seriesIndex][dataPointIndex]);
-
-					return (
-						'<div style="padding: clamp(4px, 0.83vw, 0.83vw); font-size: clamp(10px,0.83vw,0.83vw)">' +
-						'<div style="font-weight: bold;">' +
-						String(date.getDate()).padStart(2, '0') +
-						' ' +
-						monthNames[date.getMonth()] +
-						'</div>' +
-						'<div style="display: flex; gap: clamp(8px,0.83vw,0.83vw); justify-content: space-between">' +
-						'<span style="display: flex; align-items: center;">' +
-						'<div style="border-radius: 100%; height: clamp(8px,0.6vw,0.6vw); width: clamp(8px,0.6vw,0.6vw); margin-right: clamp(4px,0.24vw,0.24vw); background-color:' +
-						w.globals.colors[0] +
-						';"></div>' +
-						w.globals.initialSeries[0].name +
-						' : ' +
-						'</span>' +
-						'<span style="font-weight: bold;">' +
-						series[0][dataPointIndex].toLocaleString() +
-						'</span>' +
-						'</div>' +
-						'<div style="display: flex; gap: clamp(8px,0.83vw,0.83vw); justify-content: space-between">' +
-						'<span style="display: flex; align-items: center;">' +
-						'<div style="border-radius: 100%; height: clamp(8px,0.6vw,0.6vw); width: clamp(8px,0.6vw,0.6vw); margin-right: clamp(4px,0.24vw,0.24vw); background-color:' +
-						w.globals.colors[1] +
-						';"></div>' +
-						w.globals.initialSeries[1].name +
-						' : ' +
-						'</span>' +
-						'<span style="font-weight: bold;">' +
-						(series[1][dataPointIndex] === 0 ? 0 : -series[1][dataPointIndex]).toLocaleString() +
-						'</span>' +
-						'</div>' +
-						'</div>'
-					);
-				}
-			},
-			grid: {
-				xaxis: {
-					lines: {
-						show: false
-					}
-				},
-				yaxis: {
-					lines: {
-						show: true
+					zoom: {
+						pan: {
+							enabled: true,
+							mode: 'x',
+							threshold: 5
+						},
+						zoom: {
+							wheel: {
+								enabled: true
+							},
+							drag: {
+								enabled: false
+							},
+							pinch: {
+								enabled: true
+							},
+							mode: 'x'
+						}
 					}
 				}
 			}
-		};
-
-		// @ts-ignore
-		chart = new ApexCharts(chartElement, options);
-		chart.render();
-	});
-
-	$: if (!isLoading) {
-		chart?.updateSeries([
-			{
-				name: 'Delegated',
-				data: delegatedData,
-				color: '#0021A5'
-			},
-			{
-				name: 'Unbonded',
-				data: unbondedData,
-				color: '#099B91'
-			}
-		]);
-	}
+		});
+	};
 </script>
-
-<svelte:window bind:innerWidth />
 
 <div class="container">
 	<div class="title">Staked Per Era</div>
@@ -219,7 +123,10 @@
 			<div class="text">Unbonded</div>
 		</div>
 	</div>
-	<div class="chart" bind:this={chartElement} />
+	<ChartToolbar {chart}/>
+	<div class="chart">
+		<canvas bind:this={ctx} />
+	</div>
 </div>
 
 <style lang="postcss">
@@ -233,7 +140,7 @@
 	}
 
 	.container {
-		@apply md:h-[32vw] min-w-max;
+		@apply min-w-max;
 		@apply flex flex-col items-center justify-center;
 		@apply my-[clamp(16px,0.95vw,0.95vw)];
 	}
